move some private methods out of InstrumentContext class.
