+++ api/tests/opentrons/protocol_api/test_instrument.py
@@ -2,11 +2,8 @@
 import pytest
 from unittest import mock
 import opentrons.protocol_api as papi
-from opentrons.protocol_api.labware import Well
 from opentrons.protocols.advanced_control import transfers
-from opentrons.protocols.implementations.well import WellImplementation
-from opentrons.protocols.geometry.well_geometry import WellGeometry
-from opentrons.types import Mount, Point
+from opentrons.types import Mount
 from opentrons.protocols.api_support.types import APIVersion
 
 
@@ -121,46 +118,3 @@ def test_valid_blowout_location(
             .blow_out_strategy
 
         assert blowout_strat == expected_strat
-
-
-@pytest.mark.parametrize(argnames=["api_version", "expected_point"],
-                         argvalues=[
-                             # Above version_breakpoint:
-                             #  subtract return_height (0.5) * tip_length (1)
-                             #  from z (15)
-                             [APIVersion(2, 3), Point(10, 10, 14.5)],
-                             # Below version_breakpoint:
-                             #  add 10 to bottom (10)
-                             [APIVersion(2, 0), Point(10, 10, 20)],
-                         ])
-def test_determine_drop_target(
-        make_context_and_labware,
-        api_version,
-        expected_point):
-    fixture = make_context_and_labware(api_version)
-    lw_mock = fixture['lw1']
-    lw_mock._implementation.is_tiprack = mock.MagicMock(return_value=True)
-    lw_mock._implementation.get_tip_length = mock.MagicMock(return_value=1)
-    well = Well(
-        well_implementation=WellImplementation(
-            well_geometry=WellGeometry(
-                well_props={
-                    'shape': 'circular',
-                    'depth': 5,
-                    'totalLiquidVolume': 0,
-                    'x': 10,
-                    'y': 10,
-                    'z': 10,
-                    'diameter': 5,
-                },
-                parent_point=Point(0, 0, 0),
-                parent_object=lw_mock._implementation),
-            display_name="",
-            has_tip=False,
-            name="A1",
-        ),
-        api_level=api_version
-    )
-    r = fixture['instr']._determine_drop_target(well)
-    assert r.labware.object == well
-    assert r.point == expected_point
